feat: basic subtask relationships, reverting some refactorings
