+++ tests/project.rs
@@ -5,16 +5,17 @@
 use chrono::Utc;
 use planter_core::{
     person::{Name, Person},
-    project::Project,
+    project::{Project, TimeRelationship},
     resources::{Consumable, Material, NonConsumable, Resource},
     stakeholders::Stakeholder,
     task::Task,
-    tasks::TimeRelationship,
 };
 
 #[test]
-/// The standard workflow when creating a project involves initializing it with a name, optionally a description, and a start date.
-/// The project is kept mutable and the user can add/remove tasks, resources, stakeholders, and other relevant information.
+/// The standard workflow when creating a project involves initializing it with
+/// a name, optionally a description, and a start date. The project is kept
+/// mutable and the user can add/remove tasks, resources, stakeholders, and
+/// other relevant information.
 fn test_project() {
     // Initialize a project with a name, description, and start date.
     let start_date = Utc::now();
@@ -24,14 +25,27 @@ fn test_project() {
         )
         .with_start_date(start_date);
 
-    // Add tasks to the project
+    // Add tasks to the project.
     project.add_task(Task::new("Become world leader".to_owned()));
+    project.add_task(Task::new("Get rich".to_owned()));
+    project.add_task(Task::new("Open a proprietary software house".to_owned()));
+    project.add_task(Task::new("Prey on free software projects".to_owned()));
     project.add_task(Task::new("Profit".to_owned()));
-    assert_eq!(project.tasks().count(), 2);
+    assert_eq!(project.tasks().count(), 5);
 
-    // Add relationships between tasks of the project.
+    // Add subtask relatonships to the project.
+    project.add_subtask(1, 2);
+    project.add_subtask(1, 3);
+
+    assert_eq!(project.subtasks(1).len(), 2);
+
+    // Add time relationships between tasks of the project.
     project
-        .add_relationship(0, 1, TimeRelationship::StartToFinish)
+        .add_time_relationship(0, 1, TimeRelationship::StartToFinish)
+        .expect("Tasks to exist and circular dependencies not present");
+
+    project
+        .add_time_relationship(0, 4, TimeRelationship::StartToFinish)
         .expect("Tasks to exist and circular dependencies not present");
 
     // Add a non consumable material to the project
